+++ R/dicho.R
@@ -95,16 +95,23 @@ dicho <- function(x, ..., dich.by = "median", as.num = FALSE, var.label = NULL, 
     # do we have a grouped data frame?
     if (inherits(.dat, "grouped_df")) {
 
-      # get grouped data, as nested data frame
-      grps <- get_grouped_data(.dat)
+      # get grouping indices and variables
+      grps <- dplyr::group_indices(.dat)
+      grp.vars <- dplyr::group_vars(.dat)
+
+      # names of grouping variables
+      vars <- colnames(.dat)[colnames(.dat) %nin% grp.vars]
+      .dat <- dplyr::ungroup(.dat)
 
       # iterate all groups
-      for (i in seq_len(nrow(grps))) {
-        # get data from each single group
-        group <- grps$data[[i]]
+      for (i in unique(grps)) {
+
+        # slice cases for each group
+        keep <- which(grps == i)
+        group <- dplyr::slice(.dat, !! keep)
 
         # now iterate all variables of interest
-        for (j in colnames(group)) {
+        for (j in vars) {
           group[[j]] <- dicho_helper(
             x = group[[j]],
             dich.by = dich.by,
@@ -115,15 +122,11 @@ dicho <- function(x, ..., dich.by = "median", as.num = FALSE, var.label = NULL, 
         }
 
         # write back data
-        grps$data[[i]] <- group
+        .dat[keep, ] <- group
       }
 
-      # unnest data frame
-      x <- tidyr::unnest(grps)
-
       # remove grouping column
-      .dat <- .dat[colnames(.dat) %nin% dplyr::group_vars(.dat)]
-
+      x <- tibble::as_tibble(.dat[colnames(.dat) %nin% grp.vars])
     } else {
       # iterate variables of data frame
       for (i in colnames(.dat)) {
@@ -135,15 +138,14 @@ dicho <- function(x, ..., dich.by = "median", as.num = FALSE, var.label = NULL, 
           val.labels = val.labels
         )
       }
-    }
 
-    # coerce to tibble and select only recoded variables
-    x <- tibble::as_tibble(x[colnames(.dat)])
+      # coerce to tibble and select only recoded variables
+      x <- tibble::as_tibble(x[colnames(.dat)])
+    }
 
     # add suffix to recoded variables?
-    if (!is.null(suffix) && !sjmisc::is_empty(suffix)) {
+    if (!is.null(suffix) && !sjmisc::is_empty(suffix))
       colnames(x) <- sprintf("%s%s", colnames(x), suffix)
-    }
 
     # combine data
     if (append) x <- dplyr::bind_cols(orix, x)