+++ R/std.R
@@ -67,6 +67,10 @@ center <- function(x, ..., include.fac = TRUE, append = FALSE, suffix = "_c") {
 }
 
 
+#' @importFrom purrr map_df
+#' @importFrom tibble as_tibble
+#' @importFrom tidyr unnest
+#' @importFrom dplyr group_indices group_vars slice ungroup
 std_and_center <- function(x, .dat, include.fac, append, standardize, suffix) {
   if (is.data.frame(x)) {
 
@@ -76,16 +80,23 @@ std_and_center <- function(x, .dat, include.fac, append, standardize, suffix) {
     # do we have a grouped data frame?
     if (inherits(.dat, "grouped_df")) {
 
-      # get grouped data, as nested data frame
-      grps <- get_grouped_data(.dat)
+      # get grouping indices and variables
+      grps <- dplyr::group_indices(.dat)
+      grp.vars <- dplyr::group_vars(.dat)
+
+      # names of grouping variables
+      vars <- colnames(.dat)[colnames(.dat) %nin% grp.vars]
+      .dat <- dplyr::ungroup(.dat)
 
       # iterate all groups
-      for (i in seq_len(nrow(grps))) {
-        # get data from each single group
-        group <- grps$data[[i]]
+      for (i in unique(grps)) {
+
+        # slice cases for each group
+        keep <- which(grps == i)
+        group <- dplyr::slice(.dat, !! keep)
 
         # now iterate all variables of interest
-        for (j in colnames(group)) {
+        for (j in vars) {
           group[[j]] <- std_helper(
             x = group[[j]],
             include.fac = include.fac,
@@ -94,14 +105,11 @@ std_and_center <- function(x, .dat, include.fac, append, standardize, suffix) {
         }
 
         # write back data
-        grps$data[[i]] <- group
+        .dat[keep, ] <- group
       }
 
-      # unnest data frame
-      x <- tidyr::unnest(grps)
-
       # remove grouping column
-      .dat <- .dat[colnames(.dat) %nin% dplyr::group_vars(.dat)]
+      x <- tibble::as_tibble(.dat[colnames(.dat) %nin% grp.vars])
 
     } else {
       # iterate variables of data frame
@@ -112,16 +120,15 @@ std_and_center <- function(x, .dat, include.fac, append, standardize, suffix) {
           standardize = standardize
         )
       }
+
+      # coerce to tibble and select only recoded variables
+      x <- tibble::as_tibble(x[colnames(.dat)])
     }
 
-    # coerce to tibble and select only recoded variables
-    x <- tibble::as_tibble(x[colnames(.dat)])
-
     # add suffix to recoded variables?
-    if (!is.null(suffix) && !sjmisc::is_empty(suffix)) {
+    if (!is.null(suffix) && !sjmisc::is_empty(suffix))
       colnames(x) <- sprintf("%s%s", colnames(x), suffix)
 
-      }
     # combine data
     if (append) x <- dplyr::bind_cols(orix, x)
   } else {